+++ README.md
@@ -21,3 +21,6 @@
 
 ## 4st
 * 자동차 객체 생성. 자동차 객체는 위치와 이름을 갖는다.
+
+## 5st
+* setter 제거. 객체 생성시 생성자로 setter를 대체한다.
