--- src/main/java/Car.java
@@ -1,26 +1,0 @@
-public class Car {
-
-    private int position;
-    private String name;
-
-    public Car(int position, String name) {
-        this.position = position;
-        this.name = name;
-    }
-
-    public int getPosition() {
-        return position;
-    }
-
-    public void setPosition(int position) {
-        this.position = position;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
