+++ src/main/java/study/StringCalc.java
@@ -1,3 +1,5 @@
+package study;
+
 public class StringCalc {
 
     public static void main(String[] args) {
